+++ packages/editor/src/components/preview-dropdown/utils.js
@@ -1,7 +1,5 @@
-export function shouldShowTemplateOption( {
-	isTemplate,
-	templateId,
-	isFocusedTemplatePart,
-} ) {
-	return ! isTemplate && !! templateId && ! isFocusedTemplatePart;
+import { DESIGN_POST_TYPES } from '../../store/constants';
+
+export function shouldShowTemplateOption( { postType, templateId } ) {
+	return !! templateId && ! DESIGN_POST_TYPES.includes( postType );
 }